requirements.txt
asgiref==3.6.0
Django==4.2
django-taggit==3.1.0
sqlparse==0.4.4
